drivers/slipdev: Split out networking/coap into files
